[B] Update overlay + fix hide on redirect condition
